agregando funcion editar producto

+++ controllers/product-controller.js
@@ -7,21 +7,27 @@ const createNewLine = (image, category, name, price, description, id) => {
     const line = document.createElement("div");
     const content =
         `
-        <img class="product__line__products__details__product-img" src="${image}" alt="">
-                        <button class="product__line__products__details__delete-btn" id= ${id}><i class="fa fa-trash" aria-hidden="true"></i></button>
-                        <button class="product__line__products__details__edit-btn" id=${id}><i class='fas fa-pen'></i></button>
-                        <label class="product__line__products__details__product__category" id="${category}"></label>
-                        <h2 class="product__line__products__details__product-name">${name}</h2>
-                        <h2 class="product__line__products__details__product-price">$${price}</h2>
-                        <a class="product__line__products__details__product-btn" href="">Ver producto</a>
+        <div class = "product__line__products__details">
+            <img class="product__line__products__details__product-img" src="${image}" alt="">
+            <button type = "button" class="product__line__products__details__delete-btn" id= ${id}><i class="fa fa-trash" aria-hidden="true"></i></button>
+            <button class="product__line__products__details__edit-btn"><a href="../screens/edit-product.html?id=${id}"><i class='fas fa-pen'></i></a></button>
+            <label class="product__line__products__details__product__category" id="${category}"></label>
+            <h2 class="product__line__products__details__product-name">${name}</h2>
+            <h2 class="product__line__products__details__product-price">$${price}</h2>
+            <a class="product__line__products__details__product-btn" href="">Ver producto</a>
+        </div>
         `;
     line.innerHTML = content;
-    const btn = document.querySelector("button");
-    btn.addEventListener("click", () => {
-        const id = btn.id;
-        productServices.deleteProduct(id).then(respuesta => {
-        }).catch((error) => alert("Ocurrio un error"));
-    });
+    line.dataset.id = id;
+    const btnDelete = document.querySelector(".product__line__products__details__delete-btn");
+    if(btnDelete){
+        btnDelete.addEventListener("click", () => {
+            const id = btnDelete.id;
+            productServices.deleteProduct(id).then(respuesta => {
+            }).catch((error) => console.error("Ocurrio un error"));
+        });
+    }
+    
     return line;
 };
 const product = document.querySelector("[data-products]");
@@ -32,7 +38,6 @@ const productTypeDiversos = document.querySelector("[data-product3]");
 productServices.productsList().then((data) => {
     data.forEach(({ image, category, name, price, description, id }) => {
         const newLine = createNewLine(image, category, name,price, description, id);
-        console.log(category);
 
         const url = window.location.href;
         if (url.includes("products")) {
